+++ cookbook/attributes/default.rb
@@ -10,3 +10,6 @@ node.default['java']['jna']['platform']['url'] =
   'https://github.com/twall/jna/raw/4.1.0/dist/jna-platform.jar'
 node.default['java']['jna']['platform']['sha256sum'] =
   'f91ba7c0f26c34f04bf57d2ae30d4b19f906e7bb1de90eb3e1f4fdbf45d0c541'
+
+# Maven
+node.default['java']['maven']['enable'] = true