+++ cookbook/metadata.rb
@@ -8,5 +8,6 @@ version '0.2.0'
 
 depends 'java', '= 1.22.0'
 depends 'java_ext', '= 0.1.1'
+depends 'maven', '= 1.1.0'
 
 supports 'centos', '= 6.5'