+++ cookbook/recipes/default.rb
@@ -5,3 +5,6 @@
 
 # Configure java
 include_recipe 'cb-java::java'
+
+# Configure maven
+include_recipe 'cb-java::maven' if node['java']['maven']['enable']