+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Update release of cb-java
 - Includes JCE
 - Includes JNA
+- Include Maven
 
 # 0.1.0